numeric relations and some refactoring
